Refactored to separate concerns
